feat: add more valiidation into users routes

+++ controllers/users.controller.ts
@@ -1,5 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import prisma from "../db/prismaClient";
+import isValidString from "../utils/isValidString.util";
 import protectPasswordUtil from "../utils/protectPassword.util";
 
 const getSingleUser = async (
@@ -7,9 +8,15 @@ const getSingleUser = async (
   res: Response,
   next: NextFunction
 ) => {
-  try {
-    const { userId } = params;
+  const { userId } = params;
 
+  if (!isValidString(userId)) {
+    return res.status(400).json({
+      error: "The `taskId` query parameter is required and must not be empty.",
+    });
+  }
+
+  try {
     const user = await prisma.user.findUnique({
       where: {
         id: userId,
@@ -17,7 +24,7 @@ const getSingleUser = async (
     });
 
     if (!user) {
-      return res.status(404).json({ message: "User not found!" });
+      return res.status(404).json({ error: "User not found!" });
     }
 
     res.json({
@@ -29,9 +36,14 @@ const getSingleUser = async (
 };
 
 const updateUser = async (req: Request, res: Response, next: NextFunction) => {
-  try {
-    const { userId } = req.params;
+  const { userId } = req.params;
+  if (!isValidString(userId)) {
+    return res.status(400).json({
+      error: "The `taskId` query parameter is required and must not be empty.",
+    });
+  }
 
+  try {
     await prisma.user.update({
       where: {
         id: userId,
@@ -42,16 +54,21 @@ const updateUser = async (req: Request, res: Response, next: NextFunction) => {
       },
     });
 
-    res.json("User Updated Successfully");
+    res.json("User Updated Successfully.");
   } catch (error) {
     next(error);
   }
 };
 
 const deleteUser = async (req: Request, res: Response, next: NextFunction) => {
-  try {
-    const { userId } = req.params;
+  const { userId } = req.params;
+  if (!isValidString(userId)) {
+    return res.status(400).json({
+      error: "The `taskId` query parameter is required and must not be empty.",
+    });
+  }
 
+  try {
     await prisma.user.delete({
       where: {
         id: userId,
